Moved percent_bar in tools/
