+++ eco_perf/terminal_interface/CMakeLists.txt
@@ -1,8 +1,9 @@
 add_library(terminal_interface INTERFACE)
 
 target_link_libraries(terminal_interface INTERFACE
-    io graphics terminal)
+    io graphics terminal tools)
 
 add_subdirectory(io)
 add_subdirectory(graphics)
 add_subdirectory(terminal)
+add_subdirectory(tools)
