+++ eco_perf/terminal_interface/tools/CMakeLists.txt
@@ -1,0 +1,4 @@
+add_library(tools
+    percent_bar.h percent_bar.c
+)
+target_link_libraries(tools io m)  # math library
