+++ eco_perf/terminal_interface/tools/percent_bar.c
